+++ src/entry.cli.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-// biome-ignore lint/nursery/useImportRestrictions: This import is intentional.
-import "../dist/main.js"