+++ src/program/InvalidConfigurationProgram/InvalidConfigurationProgram.tests.ts
@@ -2,7 +2,7 @@
 import { injectFileSystemMock } from "+adapters/FileSystem/FileSystem.mock"
 import { injectLoggerMock } from "+adapters/Logger/Logger.mock"
 
-import { mainProgram } from "+program/Program"
+import { updraftProgram } from "+program/UpdraftProgram"
 import type { ExitCode } from "+utilities/ErrorUtilities"
 import { dedent } from "+utilities/StringUtilities"
 import { beforeEach, describe, expect, it } from "vitest"
@@ -35,7 +35,7 @@ describe.each`
 		let actualExitCode: ExitCode | null = null
 
 		beforeEach(async () => {
-			actualExitCode = await mainProgram(props.invalidArgs)
+			actualExitCode = await updraftProgram(props.invalidArgs)
 		})
 
 		it("returns an exit code of 2", () => {